+++ rimu-ast/Cargo.toml
@@ -1,11 +1,13 @@
 [package]
-name = "rimu-report"
+name = "rimu-ast"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-chumsky = "0.9.2"
-ariadne = "0.3.0"
-thiserror = "1.0.44"
+rust_decimal = "1.31.0"
+rimu-meta = { path = "../rimu-meta" }
+
+[dev-dependencies]
+pretty_assertions = "1.4.0"
