+++ rimu-ast/src/block.rs
@@ -1,14 +1,13 @@
-use rimu_expr::Expression;
-use rimu_report::Spanned;
+use rimu_meta::Spanned;
 
-use crate::operation::Operation;
+use crate::{expression::Expression, operation::BlockOperation};
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub enum Block {
     Object(Vec<(Spanned<String>, SpannedBlock)>),
     List(Vec<SpannedBlock>),
     Expression(Expression),
-    Operation(Box<Operation>),
+    Operation(Box<BlockOperation>),
 }
 
 pub type SpannedBlock = Spanned<Block>;
