--- rimu-block/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "rimu-block"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ariadne = "0.3.0"
-chumsky = "0.9.2"
-line-span = "0.1.5"
-rimu-report = { path = "../rimu-report" }
-rimu-token = { path = "../rimu-token" }
-rimu-expr = { path = "../rimu-expr" }
-thiserror = "1.0.44"
-
-[dev-dependencies]
-map-macro = "0.2.6"
-pretty_assertions = "1.4.0"
