--- rimu-env/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "rimu-env"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-lazy_static = "1.4.0"
-regex = "1.9.1"
-rimu-value = { path = "../rimu-value" }
-thiserror = "1.0.44"
-
-[dev-dependencies]
-pretty_assertions = "1.4.0"
