+++ rimu-parse/src/token.rs
@@ -1,7 +1,7 @@
 use rust_decimal::Decimal;
 use std::fmt;
 
-use rimu_report::Spanned;
+use rimu_meta::Spanned;
 
 /// A [`Token`] is the smallest logical unit evaluated by the compiler.
 /// It containes either an operator or a literal value.
