+++ rimu-repl/src/main.rs
@@ -1,7 +1,7 @@
 use rustyline::error::ReadlineError;
 use rustyline::{DefaultEditor, Result};
 
-use rimu::{evaluate_expr, parse_expr, Environment, ReportError, SourceId};
+use rimu::{evaluate_expression, parse_expression, Environment, ReportError, SourceId};
 
 fn main() -> Result<()> {
     // `()` can be used when no completer is required
@@ -15,7 +15,7 @@ fn main() -> Result<()> {
         match readline {
             Ok(line) => {
                 rl.add_history_entry(line.as_str())?;
-                let (expr, errors) = parse_expr(line.as_str(), SourceId::repl());
+                let (expr, errors) = parse_expression(line.as_str(), SourceId::repl());
                 if !errors.is_empty() {
                     for error in errors {
                         error.display(line.as_str(), SourceId::repl());
@@ -27,7 +27,7 @@ fn main() -> Result<()> {
                         continue;
                     };
                 // println!("Expression: {}", expr);
-                match evaluate_expr(&expr, &env) {
+                match evaluate_expression(&expr, &env) {
                     Ok(value) => println!("{}", value),
                     Err(error) => error.display(line.as_str(), SourceId::repl()),
                 }
