+++ rimu/src/lib.rs
@@ -1,6 +1,7 @@
-pub use rimu_block::{parse, Error as ParseError};
-pub use rimu_env::{Environment, EnvironmentError};
-pub use rimu_eval::{evaluate_block as evaluate, evaluate_expr, EvalError};
-pub use rimu_expr::{parse as parse_expr, Error as ParseExprError};
-pub use rimu_report::{ReportError, SourceId, Span, Spanned};
+pub use rimu_eval::{
+    evaluate_block as evaluate, evaluate_block, evaluate_expression, Environment, EnvironmentError,
+    EvalError,
+};
+pub use rimu_meta::{ReportError, SourceId, Span, Spanned};
+pub use rimu_parse::{parse_block, parse_block as parse, parse_expression, Error as ParseError};
 pub use rimu_value::{convert, from_value, List, Number, Object, Value, ValueError};
